reformat everything with PhpStorm settings (#20)
* style: reformat everything according to phpstorm

This isn't ideal for lots of reasons, so I'll replace it with phpcs fixer and a PSR12 ruleset.  which will probably cause another format apocalypse, but what can you do?

* refactor: rm all themes except twentytwentyfive and twentytwentyone

twentytwentyfive is the current theme whereas twentytwentyone is the last of the yearly themes that isn't gutenberg.  twentytwentyfive is also a bit broken at the moment, so I may re-add twentytwentyfour at a later date.

* refactor: rip out phpcs entirely

it will be replaced by phpcs fixer, but for now I'm relying on PhpStorm inspections.